config fix, farid size fix

+++ cmd/eupf/ebpf_pdr.go
@@ -8,7 +8,7 @@ import (
 
 type PdrInfo struct {
 	OuterHeaderRemoval uint8
-	FarId              uint32
+	FarId              uint16
 }
 
 func (o *BpfObjects) PutPdrUpLink(teid uint32, pdrInfo PdrInfo) error {
